create additional wheels and adjust versions
- adjust version
- add ability to create manylinux wheels using package_endpoints.txt

+++ build-linux-wheels.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -e
 
-GDAL_BUILD_PATH=/src/gdal-2.3.0/swig/python
+GDAL_BUILD_PATH=/src/gdal-2.1.2/swig/python
 ORIGINAL_PATH=$PATH
 UNREPAIRED_WHEELS=/tmp/wheels
 
@@ -12,14 +12,41 @@ source /opt/rh/devtoolset-2/enable
 pushd ${GDAL_BUILD_PATH}
 for PYBIN in /opt/python/*/bin; do
     if [[ $PYBIN == *"26"* ]]; then continue; fi
+    if [[ $PYBIN == *"27m/"* ]]; then continue; fi
     if [[ $PYBIN == *"33"* ]]; then continue; fi
+    if [[ $PYBIN == *"34"* ]]; then continue; fi
+    if [[ $PYBIN == *"35"* ]]; then continue; fi
+    if [[ $PYBIN == *"37"* ]]; then continue; fi
     export PATH=${PYBIN}:$ORIGINAL_PATH
     rm -rf build
     CFLAGS="-std=c++11" python setup.py bdist_wheel -d ${UNREPAIRED_WHEELS}
 done
 popd
 
-# Bundle GEOS into the wheels
+pushd /tmp
+filename='/io/package_endpoints.txt'
+filelines=`cat $filename`
+for line in $filelines ; do
+    curl -f -L -O $line
+    fn="${line##*/}"
+    tar xzf ${fn}
+    pushd "${fn//.tar.gz}"
+    for PYBIN in /opt/python/*/bin; do
+        if [[ $PYBIN == *"26"* ]]; then continue; fi
+        if [[ $PYBIN == *"27m/"* ]]; then continue; fi
+        if [[ $PYBIN == *"33"* ]]; then continue; fi
+        if [[ $PYBIN == *"34"* ]]; then continue; fi
+        if [[ $PYBIN == *"35"* ]]; then continue; fi
+        if [[ $PYBIN == *"37"* ]]; then continue; fi
+        export PATH=${PYBIN}:$ORIGINAL_PATH
+        rm -rf build
+        CFLAGS="-std=c++11" python setup.py bdist_wheel -d ${UNREPAIRED_WHEELS}
+    done
+    popd
+done
+popd
+
+# Bundle dependency libs into the wheels
 for whl in ${UNREPAIRED_WHEELS}/*.whl; do
     auditwheel repair ${whl} -w wheels
 done
